LayoutTests/imported/w3c/web-platform-tests/trusted-types/eval-csp-tt-default-policy-expected.txt


PASS eval of TrustedScript works.
PASS indirect eval of TrustedScript works.
FAIL eval of string works. assert_equals: expected 5 but got 2
PASS eval of !TrustedScript and !string works.
PASS Function constructor of TrustedScript works.
FAIL Function constructor of string works. assert_equals: expected 5 but got 2
